compile: fix 64-bit compilation warnings
The compiler issues warnings when converting 64-bit pointers to 32-bit
integers. Use uintptr_t instead.

+++ src/ipc/ipc-host-ptable.c
@@ -105,7 +105,7 @@ static int ipc_get_page_descriptors(struct dma *dmac, uint8_t *page_table,
 	dma_sg_init(&config.elem_array);
 
 	/* set up DMA descriptor */
-	elem.dest = (uint32_t)page_table;
+	elem.dest = (uintptr_t)page_table;
 	elem.src = ring->phy_addr;
 
 	/* source buffer size is always PAGE_SIZE bytes */
